A lot more NuGet-Packages and clean-up of binaries. The repository should not contain binaries.

+++ nuget-packages/pack.ps1
@@ -2,9 +2,9 @@
 Import-Module -Name "./Invoke-MsBuild.psm1" 
 
 
-$projects = "DatenMeister", "DatenMeister.CSV"
+$projects = "DatenMeister", "DatenMeister.Runtime", "DatenMeister.CSV", "DatenMeister.XMI", "DatenMeister.Uml"
 
-ForEach($project in $projects)
+ForEach ($project in $projects)
 {
     Write-Host "Creating package for $project"
     Push-Location
@@ -14,28 +14,38 @@ ForEach($project in $projects)
     if ( -Not $buildResult)
     {
         Write-Host "Build of $project failed. Aborting"
+        Pop-Location
         Exit
     }
 
     Pop-Location
 
-    Remove-Item -Path .\$project\lib -Recurse
+    if (Test-Path .\$project\lib)
+    {
+        Remove-Item -Path .\$project\lib -Recurse
+    }
+
     New-Item -ItemType Directory -Path $project/lib -Force | Out-Null
     New-Item -ItemType Directory -Path $project/lib/dotnet -Force | Out-Null
 
-    Copy-Item ..\src\$project\bin\Release/*.* $project/lib/dotnet 
+    Copy-Item ..\src\$project\bin\Release\$project.dll $project/lib/dotnet 
+    Copy-Item ..\src\$project\bin\Release\$project.pdb $project/lib/dotnet 
 
 
     Push-Location
 
     cd $project
-    nuget pack $project.nuspec
+    nuget pack $project.nuspec | Out-Null
 
     Pop-Location
 
     New-Item -ItemType Directory -Path All -Force | Out-Null
     Copy-Item $project/*.nupkg All/
+
+    # Now clean the created directories, only the 'All' folder shall contain values. 
+    Remove-Item $project/*.nupkg    
+    Remove-Item -Path .\$project\lib -Recurse
 }
 
-
+Write-Host
 Write-Host "The packages are created, you can now push them via nuget push All/yourpackage.nupkg"
